Update Shadows for new World
This updates the shadows to work in an arbitrary world. It grabs the
height from the terrain and uses this for the Near/Far planes. The X/Y
boundaries are centered around the player and offset by a constant
distance.

Along with this it adds the code for increasing bias on edges with
tangents perpindicular to the light and bumps up the resolution of the
shadow maps through an added constant.