Particles: The whole thing
My bad on not splitting this up into commits. This right now sets up
particle systems as components that can be added through the json file.

The behavior of the particle systems are pretty specific to the engine
trail but could potentially be pulled out into the json if we want
different types of particles. I didn't add that right now because I
don't know if we will actually use particles except for fire like
effects and there's no need for abstraction that we won't use.

The one engine trail specific thing I haven't figured out yet is how to
specify the condition for when to create particles. Since this is
essentially a function and we can't write the function in json.